% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIIV_Functions.R
\name{CIIV.TestingSelectionEstimation}
\alias{CIIV.TestingSelectionEstimation}
\title{Internal CIIV Functions}
\usage{
CIIV.TestingSelectionEstimation(
  Y,
  D,
  Z,
  U,
  Covariates,
  Exogenous,
  y,
  z,
  CovariatesNames,
  InstrumentNames,
  betaIV,
  sdIV,
  alpha = 0.05,
  tuning = 0.1/log(length(Y)),
  robust = TRUE,
  gamma_D,
  Covmat_gamma,
  firststage = FALSE,
  relevant_instruments,
  Nr_relevant
)
}
\description{
These are not to be called by the user.
}
